Changing soft fields for Image object
Till now there was an idea to expose full path to image file and thumbnail.
It's quite impossible in certain situations, where path to the media files is
not known. Therefore 'imagepath' was removed due to redundancy with filename
stored in db, and 'thumpath', which now is a simple calculation of the
expected thumbnail relative filepath.